requirements.txt
torch==2.11.0
torchvision==0.26.0
timm==1.0.26
numpy==2.4.4
pandas==3.0.2
scikit-learn==1.8.0
pydicom==3.0.2
opencv-python==4.13.0.92
Pillow==12.2.0
matplotlib==3.10.8
seaborn==0.13.2
streamlit==1.56.0
grad-cam==1.5.5
albumentations==2.0.8
tqdm==4.67.3
fastapi==0.115.12
uvicorn==0.34.3
python-multipart==0.0.20
reportlab==4.2.5
